+++ src/main/java/com/transferwise/openbanking/client/api/registration/domain/ApplicationType.java
@@ -1,0 +1,17 @@
+package com.transferwise.openbanking.client.api.registration.domain;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum ApplicationType {
+    WEB("web"),
+    MOBILE("mobile");
+
+    private final String value;
+
+    @JsonValue
+    public String getValue() {
+        return value;
+    }
+}